pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tieh"
version = "0.1.0"
description = "Tiè! Send files to PC"
authors = [{ name = "Fubeeo" }]
license = { file = "LICENSE" }
readme = "README.md"
dependencies = [
    "flask",
    "flask_socketio",
    "appdirs",
    "Pillow",
    "qrcode",
]

[project.scripts]
tieh = "tieh:main"

[tool.hatch.envs.default]
dependencies = ["pre-commit"]
